move folder creation from config to plugin ctrl
